+++ components/commitdiff/commitlinediff.js
@@ -13,6 +13,7 @@ var CommitLineDiff = function(args, fileLineDiff) {
   this.server = args.server;
   this.sha1 = args.sha1;
   this.textDiffType = args.textDiffType;
+  this.wordWrap = args.wordWrap;
   this.specificDiff = ko.observable(this.getSpecificDiff());
 };
 exports.CommitLineDiff = CommitLineDiff;
@@ -24,7 +25,8 @@ CommitLineDiff.prototype.getSpecificDiff = function() {
     server: this.server,
     sha1: this.sha1,
     textDiffType: this.textDiffType,
-    isShowingDiffs: this.isShowingDiffs
+    isShowingDiffs: this.isShowingDiffs,
+    wordWrap: this.wordWrap
   });
 }