+++ CHANGELOG.md
@@ -6,7 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ## [Unreleased](https://github.com/FredrikNoren/ungit/compare/v0.10.1...master)
 
 ### Added
+- Added bare repo support [#177](https://github.com/FredrikNoren/ungit/issues/177) [#728](https://github.com/FredrikNoren/ungit/issues/728)
 - Added support for cherry-pick conflict[#701](https://github.com/FredrikNoren/ungit/issues/701)
+- Added wordwrap support for diffs [#721](https://github.com/FredrikNoren/ungit/issues/721)
 
 ### Fixed
 - Fix maxConcurrentGitOperations not limiting git processes [#707](https://github.com/FredrikNoren/ungit/issues/707)